requirements.txt
kafka-python==2.0.2
pydantic==2.5.0
python-dotenv==1.0.0
